:sparkles: CardListContent, CardList 컴포넌트 분리
