+++ src/app/(root)/(routes)/cards/components/card-list-content/index.tsx
@@ -1,0 +1,3 @@
+import CardListContent from './CardListContent'
+
+export default CardListContent
