+++ src/app/(root)/(routes)/cards/page.tsx
@@ -1,6 +1,6 @@
 import { FunctionComponent } from 'react'
 import PageTitle from '@/components/domain/page-title'
-import CardList from './components/card-list'
+import CardListContent from './components/card-list-content'
 
 interface CardListPageProps {}
 
@@ -8,7 +8,7 @@ const CardListPage: FunctionComponent<CardListPageProps> = ({}) => {
   return (
     <div>
       <PageTitle title="물건 목록" />
-      <CardList />
+      <CardListContent />
     </div>
   )
 }
